Lab-2

+++ Lab-2/types.h
@@ -1,10 +1,9 @@
 #ifndef TYPES_H
 #define TYPES_H
 
-/* 裸机编程没有标准头文件，需要自己定义基础类型 */
-typedef unsigned char      uint8;
-typedef unsigned int       uint32;
-typedef unsigned long      uint64;
-typedef long               int64;
+typedef unsigned char          uint8;
+typedef unsigned int           uint32;
+typedef unsigned long long     uint64;
+typedef long                   int64;
 
 #endif
